clang format

+++ robotics/ocs2_pinocchio/ocs2_self_collision/src/SelfCollision.cpp
@@ -37,18 +37,22 @@
 
 #include <ocs2_self_collision/SelfCollision.h>
 
-namespace ocs2 {
+namespace ocs2
+{
     SelfCollision::SelfCollision(PinocchioGeometryInterface pinocchioGeometryInterface, scalar_t minimumDistance)
-        : pinocchioGeometryInterface_(std::move(pinocchioGeometryInterface)), minimumDistance_(minimumDistance) {
+        : pinocchioGeometryInterface_(std::move(pinocchioGeometryInterface)), minimumDistance_(minimumDistance)
+    {
     }
 
 
-    vector_t SelfCollision::getValue(const PinocchioInterface &pinocchioInterface) const {
+    vector_t SelfCollision::getValue(const PinocchioInterface& pinocchioInterface) const
+    {
         const std::vector<hpp::fcl::DistanceResult> distanceArray = pinocchioGeometryInterface_.computeDistances(
             pinocchioInterface);
 
         vector_t violations = vector_t::Zero(distanceArray.size());
-        for (size_t i = 0; i < distanceArray.size(); ++i) {
+        for (int i = 0; i < distanceArray.size(); ++i)
+        {
             violations[i] = distanceArray[i].min_distance - minimumDistance_;
         }
 
@@ -57,25 +61,27 @@ namespace ocs2 {
 
 
     std::pair<vector_t, matrix_t> SelfCollision::getLinearApproximation(
-        const PinocchioInterface &pinocchioInterface) const {
+        const PinocchioInterface& pinocchioInterface) const
+    {
         const std::vector<hpp::fcl::DistanceResult> distanceArray = pinocchioGeometryInterface_.computeDistances(
             pinocchioInterface);
 
-        const auto &model = pinocchioInterface.getModel();
-        const auto &data = pinocchioInterface.getData();
+        const auto& model = pinocchioInterface.getModel();
+        const auto& data = pinocchioInterface.getData();
 
-        const auto &geometryModel = pinocchioGeometryInterface_.getGeometryModel();
+        const auto& geometryModel = pinocchioGeometryInterface_.getGeometryModel();
 
         vector_t f(distanceArray.size());
         matrix_t dfdq(distanceArray.size(), model.nq);
-        for (size_t i = 0; i < distanceArray.size(); ++i) {
+        for (int i = 0; i < distanceArray.size(); ++i)
+        {
             // Distance violation
             f[i] = distanceArray[i].min_distance - minimumDistance_;
 
             // Jacobian calculation
-            const auto &collisionPair = geometryModel.collisionPairs[i];
-            const auto &joint1 = geometryModel.geometryObjects[collisionPair.first].parentJoint;
-            const auto &joint2 = geometryModel.geometryObjects[collisionPair.second].parentJoint;
+            const auto& collisionPair = geometryModel.collisionPairs[i];
+            const auto& joint1 = geometryModel.geometryObjects[collisionPair.first].parentJoint;
+            const auto& joint2 = geometryModel.geometryObjects[collisionPair.second].parentJoint;
 
             // We need to get the jacobian of the point on the first object; use the joint jacobian translated to the point
             const vector3_t joint1Position = data.oMi[joint1].translation();
@@ -87,16 +93,16 @@ namespace ocs2 {
             // [ position jacobian ]
             // [ rotation jacobian ]
             const matrix_t pt1Jacobian = joint1Jacobian.topRows(3) - skewSymmetricMatrix(pt1Offset) * joint1Jacobian.
-                                         bottomRows(3);
+                bottomRows(3);
 
             // We need to get the jacobian of the point on the second object; use the joint jacobian translated to the point
             const vector3_t joint2Position = data.oMi[joint2].translation();
             const vector3_t pt2Offset = distanceArray[i].nearest_points[1] - joint2Position;
             matrix_t joint2Jacobian = matrix_t::Zero(6, model.nv);
             getJointJacobian(model, data, joint2, pinocchio::ReferenceFrame::LOCAL_WORLD_ALIGNED,
-                                        joint2Jacobian);
+                             joint2Jacobian);
             const matrix_t pt2Jacobian = joint2Jacobian.topRows(3) - skewSymmetricMatrix(pt2Offset) * joint2Jacobian.
-                                         bottomRows(3);
+                bottomRows(3);
 
             // To get the (approximate) jacobian of the distance, get the difference between the two nearest point jacobians, then multiply by the
             // vector from point to point
@@ -104,9 +110,9 @@ namespace ocs2 {
             // TODO(perry): is there a way to calculate a correct jacobian for the case of distanceVector = 0?
             const vector3_t distanceVector = distanceArray[i].min_distance > 0
                                                  ? (distanceArray[i].nearest_points[1] - distanceArray[i].nearest_points
-                                                    [0]).normalized()
+                                                     [0]).normalized()
                                                  : (distanceArray[i].nearest_points[0] - distanceArray[i].nearest_points
-                                                    [1]).normalized();
+                                                     [1]).normalized();
             dfdq.row(i).noalias() = distanceVector.transpose() * differenceJacobian;
         } // end of i loop
